+++ NEWS.md
@@ -3,3 +3,4 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Fixed a bug in the way that RHS constraints were being constructed if there was a variable in 
 the RHS
+* Added data quality checks to make sure that input data has no NA values